+++ src/main/java/baseball/Application.java
@@ -2,6 +2,9 @@ package baseball;
 
 import java.util.ArrayList;
 
+import baseball.domain.BaseBallGame;
+import baseball.domain.Computer;
+
 public class Application {
     public static void main(String[] args) {
         BaseBallGame baseBallGame = new BaseBallGame();
