+++ src/test/java/baseball/ComputerTest.java
@@ -1,11 +1,12 @@
 package baseball;
 
-import org.junit.jupiter.api.Test;
-
 import java.util.ArrayList;
 import java.util.HashSet;
 
+import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
+
+import baseball.domain.Computer;
 
 class ComputerTest {
